requirements.txt
APScheduler==3.9.1
async-timeout==4.0.2
boto3==1.21.6
botocore==1.24.6
Brotli==1.0.9
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cvxopt==1.2.7
cvxpy==1.1.18
cycler==0.11.0
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
ecos==2.0.10
Flask==2.0.3
Flask-Assets==2.0
Flask-Caching==1.10.1
Flask-Compress==1.12
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
fonttools==4.29.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.8.0
MarkupSafe==2.1.0
matplotlib==3.5.1
multitasking==0.0.10
net-tools==0.1.2
numpy==1.22.2
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.1
pandas-datareader==0.10.0
Pillow==9.0.1
plotly==5.6.0
psycopg2-binary==2.9.3
pyparsing==3.0.7
pyportfolioopt==1.5.1
python-binary-memcached==0.31.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
qdldl==0.1.5.post0
redis==4.3.4
requests==2.27.1
s3transfer==0.5.1
scikit-learn==1.0.2
scipy==1.8.0
scs==3.2.0
six==1.16.0
SQLAlchemy==1.4.31
tenacity==8.0.1
threadpoolctl==3.1.0
tzdata==2021.5
tzlocal==4.1
uhashring==2.1
urllib3==1.26.8
waitress==2.0.0
webassets==2.0
Werkzeug==2.0.3
wrapt==1.14.1
yfinance==0.1.70
